# specs/002-web-evolution/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Evolution Todo API
  description: |
    RESTful API for the Evolution Todo web application with multi-tenant user isolation.

    ## Authentication
    All endpoints (except health check) require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## The Identity Law
    Every task operation is filtered by the authenticated user's `user_id`. No user can access another user's data.

    ## Base URLs
    - **Development**: http://localhost:8000
    - **Production**: https://api.evolution-todo.com
  version: 2.0.0
  contact:
    name: Evolution Todo Team
    email: support@evolution-todo.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.evolution-todo.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Tasks
    description: Task CRUD operations

security:
  - BearerAuth: []

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the API is running
      security: []  # No authentication required
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-01-01T12:00:00Z"

  /api/{user_id}/tasks:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'

    get:
      tags:
        - Tasks
      summary: Get all tasks for authenticated user
      description: |
        Retrieve all tasks belonging to the authenticated user.
        The `user_id` in the URL must match the authenticated user's ID.
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: |
        Create a new task for the authenticated user.
        The task will be automatically linked to the authenticated user's ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
      - $ref: '#/components/parameters/TaskIdParam'

    put:
      tags:
        - Tasks
      summary: Update an existing task
      description: |
        Update the title and/or description of an existing task.
        The task must belong to the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: |
        Delete an existing task.
        The task must belong to the authenticated user.
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
      - $ref: '#/components/parameters/TaskIdParam'

    patch:
      tags:
        - Tasks
      summary: Toggle task completion status
      description: |
        Toggle the completion status of a task.
        If the task is incomplete, it will be marked as complete, and vice versa.
        The task must belong to the authenticated user.
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Better Auth.
        Include in the Authorization header as: `Bearer <token>`

  parameters:
    UserIdParam:
      name: user_id
      in: path
      required: true
      description: The authenticated user's ID (must match JWT user_id)
      schema:
        type: integer
        minimum: 1
        example: 123

    TaskIdParam:
      name: task_id
      in: path
      required: true
      description: The task ID
      schema:
        type: integer
        minimum: 1
        example: 42

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (cannot be empty or whitespace-only)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title
          example: "Buy groceries (updated)"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Updated task description
          example: "Milk, eggs, bread, cheese"

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 42
        user_id:
          type: integer
          description: Owner's user ID
          example: 123
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Task description (optional)
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Task completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2026-01-01T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-01-01T10:30:00Z"

    TasksResponse:
      type: object
      required:
        - tasks
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: List of all tasks for the user
          example:
            - id: 1
              user_id: 123
              title: "Buy groceries"
              description: "Milk, eggs, bread"
              completed: false
              created_at: "2026-01-01T10:00:00Z"
              updated_at: "2026-01-01T10:00:00Z"
            - id: 2
              user_id: 123
              title: "Write report"
              description: null
              completed: true
              created_at: "2026-01-01T11:00:00Z"
              updated_at: "2026-01-01T12:00:00Z"

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

  responses:
    BadRequestError:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            emptyTitle:
              summary: Empty title
              value:
                detail: "Title cannot be empty"
            titleTooLong:
              summary: Title too long
              value:
                detail: "Title must be 255 characters or less"

    UnauthorizedError:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing token
              value:
                detail: "Unauthorized"
            invalidToken:
              summary: Invalid token
              value:
                detail: "Invalid token"
            expiredToken:
              summary: Expired token
              value:
                detail: "Token expired"

    ForbiddenError:
      description: Forbidden - authenticated user does not match URL user_id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Forbidden"

    NotFoundError:
      description: Not found - task does not exist or does not belong to user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"